test: try catch를 사용하지 않고 예외를 전파하도록 변경
