+++ src/test/java/nextstep/subway/acceptance/station/StationAcceptanceTest.java
@@ -1,35 +1,19 @@
 package nextstep.subway.acceptance.station;
 
-import io.restassured.RestAssured;
+import com.fasterxml.jackson.core.JsonProcessingException;
 import io.restassured.response.ExtractableResponse;
 import io.restassured.response.Response;
-import org.junit.jupiter.api.BeforeEach;
+import nextstep.subway.acceptance.AcceptanceTest;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
-import org.springframework.boot.test.context.SpringBootTest;
-import org.springframework.boot.web.server.LocalServerPort;
 import org.springframework.http.HttpStatus;
-import org.springframework.test.context.jdbc.Sql;
 
 import java.util.List;
 
-import static nextstep.subway.acceptance.station.StationPrepare.지하철역_생성_요청;
-import static nextstep.subway.acceptance.station.StationPrepare.지하철역_조회_요청;
-import static nextstep.subway.acceptance.station.StationPrepare.지하철역_삭제_요청;
+import static nextstep.subway.acceptance.station.StationAcceptanceStep.*;
 import static org.assertj.core.api.Assertions.assertThat;
 
-@Sql("classpath:/database-init.sql")
-@DisplayName("지하철역 관련 기능")
-@SpringBootTest(webEnvironment = SpringBootTest.WebEnvironment.RANDOM_PORT)
-class StationAcceptanceTest {
-    @LocalServerPort
-    int port;
-
-    @BeforeEach
-    public void setUp() {
-        RestAssured.port = port;
-    }
-
+class StationAcceptanceTest extends AcceptanceTest {
     /**
      * When 지하철역을 생성하면
      * Then 지하철역이 생성된다
@@ -37,7 +21,7 @@ class StationAcceptanceTest {
      */
     @DisplayName("지하철역을 생성한다.")
     @Test
-    void createStation() {
+    void createStation() throws JsonProcessingException {
         // when
         ExtractableResponse<Response> response = 지하철역_생성_요청("강남역");
 
@@ -45,7 +29,7 @@ class StationAcceptanceTest {
         assertThat(response.statusCode()).isEqualTo(HttpStatus.CREATED.value());
 
         // then
-        List<String> stationNames = 지하철역_조회_요청();
+        List<String> stationNames = 지하철역_조회_결과();
         assertThat(stationNames).containsAnyOf("강남역");
     }
 
@@ -57,11 +41,11 @@ class StationAcceptanceTest {
     // TODO: 지하철역 목록 조회 인수 테스트 메서드 생성
     @DisplayName("지하철역 목록을 조회한다.")
     @Test
-    void getStations() {
+    void getStations() throws JsonProcessingException {
         지하철역_생성_요청("잠실역");
         지하철역_생성_요청("선릉역");
 
-        List<String> 지하철역_목록 = 지하철역_조회_요청();
+        List<String> 지하철역_목록 = 지하철역_조회_결과();
 
         assertThat(지하철역_목록).containsExactly("잠실역", "선릉역");
         assertThat(지하철역_목록).hasSize(2);
@@ -75,12 +59,12 @@ class StationAcceptanceTest {
     // TODO: 지하철역 제거 인수 테스트 메서드 생성
     @DisplayName("지하철역을 삭제한다.")
     @Test
-    void deleteStation() {
+    void deleteStation() throws JsonProcessingException {
         ExtractableResponse<Response> 교대역 = 지하철역_생성_요청("교대역");
 
         지하철역_삭제_요청(교대역.header("Location"));
 
-        List<String> 지하철역_목록 = 지하철역_조회_요청();
+        List<String> 지하철역_목록 = 지하철역_조회_결과();
         assertThat(지하철역_목록).doesNotContain("교대역");
     }
 }
